Refactor hash scroll to hook
