+++ utils/hashScroll.ts
@@ -1,0 +1,32 @@
+import { useEffect } from 'react'
+
+export const useHashScroll = () => {
+  useEffect(() => {
+    const handleHashChange = () => {
+      const hash = window.location.hash.replace('#', '')
+      if (hash) {
+        scrollToElement(hash)
+      }
+    }
+    handleHashChange()
+    window.addEventListener('hashchange', handleHashChange)
+    return () => {
+      window.removeEventListener('hashchange', handleHashChange)
+    }
+  }, [])
+}
+
+export const scrollToElement = (id: string, offset: number = 70) => {
+  const element = document.getElementById(id)
+  if (element) {
+    const elementPosition = element.getBoundingClientRect().top + window.scrollY
+
+    // offset because of the sticky navbar
+    const offsetPosition = elementPosition - offset
+
+    window.scrollTo({
+      top: offsetPosition,
+      behavior: 'smooth',
+    })
+  }
+}
